sdfsdfsd_int32

+++ JunwooSession.py
@@ -13,7 +13,7 @@ n_visible = 1
 n_hidden = 5
 
 W   = tf.Variable(tf.zeros([n_visible, n_hidden]), name="W")
-x  = tf.placeholder(tf.float32, [2, n_visible])
+x  = tf.placeholder(tf.int32, [2, n_visible])
 
 bh  = tf.Variable([[1,2,3,4,5]])
